Fix race condition in ImagePlotWidget::display
RawImageWidget::setImage is being called on a different thread than the Qt event loop,
which RawImageWidget::paintEvent is called on.  This caused a race condition since
these functions both set and read the image.  Switching to calling on the same thread
fixes both sporadic crashes in paintEvent and setImage and image freezing when displayed.